Make pandas a non-necessary requirement (#609)
