+++ nevergrad/benchmark/test_core.py
@@ -17,6 +17,7 @@ import matplotlib
 from nevergrad.optimization import optimizerlib
 from nevergrad.parametrization.utils import CommandFunction, FailedJobError
 from nevergrad.common import testing
+from . import utils
 from . import core
 from .test_xpbase import DESCRIPTION_KEYS
 matplotlib.use('Agg')
@@ -35,7 +36,7 @@ def test_moduler(value: int, expected: bool) -> None:
 
 def test_compute() -> None:
     output = core.compute("basic")
-    assert isinstance(output, core.tools.Selector)
+    assert isinstance(output, utils.Selector)
 
 
 def test_commandline_launch() -> None:
@@ -47,7 +48,7 @@ def test_commandline_launch() -> None:
                                      "--cap_index", "2", "--num_workers", "2", "--output", str(output),
                                      "--imports", str(Path(__file__).parent / "additional" / "example.py")])()
         assert output.exists()
-        df = core.tools.Selector.read_csv(str(output))
+        df = utils.Selector.read_csv(str(output))
         testing.assert_set_equal(df.columns, DESCRIPTION_KEYS | {"offset"})  # "offset" comes from the custom function
         np.testing.assert_equal(len(df), 2)
 
@@ -59,9 +60,9 @@ def test_launch() -> None:
             # commandline test
             repeated_launch("repeated_basic", cap_index=4, num_workers=2, output=output, plot=True)
             assert output.exists()
-            df = core.tools.Selector.read_csv(str(output))
+            df = utils.Selector.read_csv(str(output))
             testing.assert_set_equal(df.unique("optimizer_name"), {"DifferentialEvolution()", "OnePlusOne"})
-            assert isinstance(df, core.tools.Selector)
+            assert isinstance(df, utils.Selector)
             np.testing.assert_equal(len(df), 4)
