+++ nevergrad/benchmark/test_experiments.py
@@ -5,25 +5,25 @@
 
 import inspect
 import itertools
-from typing import Callable, Iterator, Any
+import typing as tp
 import numpy as np
 from ..functions.mlda import datasets
 from ..functions import rl
 from ..common import testing
-from ..common.tools import Selector
 from .xpbase import Experiment
+from .utils import Selector
 from . import experiments
 
 
 @testing.parametrized(**{name: (name, maker) for name, maker in experiments.registry.items()})
-def test_experiments_registry(name: str, maker: Callable[[], Iterator[experiments.Experiment]]) -> None:
+def test_experiments_registry(name: str, maker: tp.Callable[[], tp.Iterator[experiments.Experiment]]) -> None:
     with datasets.mocked_data():  # mock mlda data that should be downloaded
         check_maker(maker)  # this is to extract the function for reuse if other external packages need it
     if name not in {"realworld_oneshot", "mlda", "mldaas", "realworld"}:
         check_seedable(maker)  # this is a basic test on first elements, do not fully rely on it
 
 
-def check_maker(maker: Callable[[], Iterator[experiments.Experiment]]) -> None:
+def check_maker(maker: tp.Callable[[], tp.Iterator[experiments.Experiment]]) -> None:
     generators = [maker() for _ in range(2)]
     # check 1 sample
     sample = next(maker())
@@ -42,7 +42,7 @@ def check_maker(maker: Callable[[], Iterator[experiments.Experiment]]) -> None:
             )
 
 
-def check_seedable(maker: Any) -> None:
+def check_seedable(maker: tp.Any) -> None:
     """Randomized check of seedability for 8 first elements
     This test does not prove the complete seedability of the generator!  (would be way too slow)
     """
