+++ nevergrad/benchmark/test_plotting.py
@@ -10,8 +10,8 @@ from typing import List
 import numpy as np
 import pandas as pd
 import matplotlib
-from ..common import tools
-from ..common import testing
+from nevergrad.common import testing
+from . import utils
 matplotlib.use('Agg')
 
 
@@ -62,7 +62,7 @@ def test_create_plots_from_csv_mocked() -> None:
 
 
 def test_fight_plotter() -> None:
-    df = tools.Selector.read_csv(Path(__file__).parent / "sphere_perf_example.csv").select(
+    df = utils.Selector.read_csv(Path(__file__).parent / "sphere_perf_example.csv").select(
         optimizer_name=["OnePlusOneOptimizer", "HaltonSearch", "Powell"])
     winrates = plotting.FightPlotter.winrates_from_selection(df, ["noise_level", "budget"])
     # check data
@@ -77,7 +77,7 @@ def test_fight_plotter() -> None:
 
 def test_xp_plotter() -> None:
     opt = "OnePlusOneOptimizer"
-    df = tools.Selector.read_csv(Path(__file__).parent / "sphere_perf_example.csv").select(optimizer_name=[opt])
+    df = utils.Selector.read_csv(Path(__file__).parent / "sphere_perf_example.csv").select(optimizer_name=[opt])
     data = plotting.XpPlotter.make_data(df)
     # check data
     testing.assert_set_equal(data.keys(), {opt})
@@ -103,7 +103,7 @@ def test_remove_errors() -> None:
     np.testing.assert_array_equal(output.columns, expected.columns)
     np.testing.assert_array_equal(output.index, expected.index)
     np.testing.assert_array_equal(output, expected)
-    assert isinstance(output, plotting.tools.Selector)
+    assert isinstance(output, plotting.utils.Selector)
 
 
 def test_make_style_generator() -> None:
@@ -147,8 +147,8 @@ def test_merge_parametrization_and_optimizer() -> None:
         columns=["optimizer_name", "parametrization", "val"],
         data=[["o1", "p1", 1], ["o1", "p2", 2], ["o2", "p1", 3]]
     )
-    out = plotting.merge_parametrization_and_optimizer(tools.Selector(df))
-    assert isinstance(out, tools.Selector)
+    out = plotting.merge_parametrization_and_optimizer(utils.Selector(df))
+    assert isinstance(out, utils.Selector)
     assert out["optimizer_name"].tolist() == ["o1,p1", "o1,p2", "o2"]
     assert out["val"].tolist() == [1, 2, 3]
