+++ requirements/bench.txt
@@ -6,4 +6,5 @@ gym>=0.12.1
 torch>=1.2.0
 hiplot
 fcmaes>=0.9.5.6
+pandas>=0.23.4
 tqdm
