+++ requirements/dev.txt
@@ -8,4 +8,5 @@ sphinx_rtd_theme>=0.4.3
 recommonmark>=0.5.0
 twine>=3.1.1  # for upload to pypi
 autodocsumm>=0.1.11
+pandas>=0.23.4
 # need to install manually numpy-stubs@git+https://github.com/numpy/numpy-stubs@master
