+++ requirements/main.txt
@@ -1,4 +1,3 @@
-pandas>=0.23.4
 numpy>=1.15.0
 cma>=2.6.0
 bayesian-optimization>=1.1.0
